requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.26.4
Pillow==10.2.0
realesrgan==0.3.0
basicsr==1.4.2
